Strip lean path info from runtime exceptions
